feat: VendingMachineCoinManager 도메인 기능 구현
